Refector for doxygen

+++ interfaces/game/singles/IGetSceneGuiHeight.java
@@ -1,6 +1,6 @@
 package com.github.a2g.core.interfaces.game.singles;
 
 public interface IGetSceneGuiHeight {
-	int getSceneGuiHeight();
+	double getSceneGuiHeight();
 
 }
